// src/types/types.ts
export default interface ILabResult {
  CLINIC_NO: string;
  BARCODE: string;
  PATIENT_ID: string;
  PATIENT_NAME: string;
  DOB: string;
  GENDER: string;
  COLLECTIONDATE: string;
  COLLECTIONTIME: string;
  TESTCODE: string;
  TESTNAME: string;
  RESULT: string;
  UNIT: string;
  REFRANGELOW: string;
  REFRANGEHIGH: string;
  NOTE: string;
  NONSPECREFS: string;
}